Add workflow for building docker image
This workflow should build and push a new docker image whenever a new release is available in the upstream repo.

This is my first attempt at using github Actions so this will most likely not work on the first commit.

And since I can't test my workflows on other branches the main branch will be littered with my failed attempts.