Update build-docker-image.yml
Looks like we finally got there.

The job will now skip build step if release already exists on Docker hub.

The job runs once every day to check for new releases since I can't seem to find a way to trigger on a new release in a repo I don't own.

Final test would be to wait for a new release and see if this actually works.