Added PUID/PGID support
Since a lot of users are accustomed to these environment variables it was time
to add support for these.
When these are used the container will start as root but the application
will run as the specified UID/GID. One limitation is that one cannot use
IDs that conflict with IDs that exist in the image. That is why PUID must
be > 100 and PGID > 99.

The old way of starting the application, with `--user`, is still
supported. Using that approach one can start a container with any user
that is not root.