Don't use `/run/.containerenv` in podman check
The `/run/.containerenv` file is created if an image is built with
podman, not only when a container is started. This irritating,
undocumented bug makes it so that we can't rely on that file to indicate
if a container is started by podman.

The only thing we can check, due to this bug, is if env.var
$container==podman. For recent versions that seems to be working.
For versions of podman where this env.var. isn't pre-populated it would
be and easy fix to just add it when starting the container and you will
get podman support.